refactor: address CR notes
